requirements.txt
# --------- pytorch --------- #
torch>=2.0.0
torchvision>=0.15.0
lightning>=2.0.0
torchmetrics>=0.11.4

# --------- core dependencies --------- #
numpy>=2.0.0  # Required by datatrove
pandas>=2.0.0  # Compatible with numpy 2.x
regex>=2022.1.18  # High-performance regex with possessive quantifiers for 10M+ documents

# --------- model & training --------- #
transformers>=4.30.0  # For T5 model support
datasets>=2.12.0     # HuggingFace datasets
accelerate>=0.20.0   # For distributed training
deepspeed>=0.9.0    # For efficient training
tensorboard>=2.13.0  # For training monitoring

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml
# aim>=3.16.2  # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550

# --------- others --------- #
rootutils       # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
# sh            # for running bash commands in some tests (optional)

# --------- datatrove dependencies --------- #
# Note: DataTrove modules WordStats/SentenceStats/LangStats require NumPy < 2.0 
# due to spaCy → CuPy compatibility issues
numba           # required for librosa compatibility with NumPy 1.x
librosa         # audio processing (torchmetrics dependency)

# DataTrove stats module dependencies
# install datatrove also as pip package
datatrove
spacy>=3.0.0    # required for WordStats, SentenceStats, WordsContaminationStats (word tokenization)
fasttext-numpy2-wheel==0.9.2  # Specific version for numpy compatibility
kenlm           # required for CCNetPerplexityStats (language model perplexity calculation)
sentencepiece   # required for CCNetPerplexityStats (SentencePiece tokenization)
tokenizers      # required for TokenStats (tokenization and counting)
tldextract      # required for BaseStats (extracting top-level domains from URLs)

# --------- CUDA dependencies --------- #
# Note: RAPIDS (cudf) 23.12 is the last version supporting Python 3.10
# cudf==23.12.*
# dask-cuda==23.12.*
# dask-cudf==23.12.*

python-dotenv
pyarrow>=14.0.0
